packages/platform-auth/pyproject.toml
[project]
name = "platform-auth"
version = "0.1.0"
description = "Shared JWT auth helpers for the LOJINEXT multi-service platform"
requires-python = ">=3.12"
dependencies = [
    "PyJWT>=2.10.1",
    "cryptography>=44.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
